requirements.txt
streamlit
pandas
python-dotenv
pytz
supabase
requests
PyGithub
matplotlib




